refactor: 카드 컴포넌트 경로를 도메인별로 정리
