+++ my-app/src/components/card/Card/Card.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import Card from "./Card";
 
 const meta = {
-  title: "Components/Card",
+  title: "Components/Card/Card",
   component: Card,
   tags: ["autodocs"],
   parameters: {
