+++ my-app/src/components/common/InputSectionLayout/InputSectionLayout.tsx
