+++ my-app/src/components/common/ValidatedInputGroup/ValidatedInputGroup.stories.tsx
@@ -64,7 +64,7 @@ const InteractiveValidatedInputGroup = (args: ValidatedInputGroupProps) => {
 };
 
 const meta = {
-  title: "Components/ValidatedInputGroup",
+  title: "Components/Common/ValidatedInputGroup",
   component: ValidatedInputGroup,
   tags: ["autodocs"],
   parameters: {
