+++ my-app/src/utils/cardCompletion.ts
